crop guard/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
python-dotenv>=1.0.0
pyserial>=3.5
anthropic>=0.25.0
opencv-python>=4.9.0
numpy>=1.26.0
# TFLite — install ONE of these depending on your platform:
# tflite-runtime>=2.14.0   ← preferred (smaller, faster)
# tensorflow>=2.15.0       ← fallback if tflite-runtime unavailable
